add jobFunctionList
